SMIC/requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.2
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
pillow>=9.0.0
transformers>=4.30.0
torch>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0